requirements.txt
pmaw==3.0.0
psycopg2_binary==2.9.5
pydantic==1.10.4
